workqueue: Split out workqueue_types.h
commit b2fa8443db320c4873feca2588b957439e350890 upstream.

More sched.h dependency culling - this lets us kill a rhashtable-types.h
dependency on workqueue.h.
